Split shortcode news and memento to many files
